+++ test/helpers/omniauth_helper.rb
@@ -28,13 +28,3 @@ def dropbox_old
   }
   OmniAuth.config.add_mock(:dropbox_oauth2, omniauth_hash_old)
 end
-
-def log_in_user(user_id)
-  session[:user] = user_id
-end
-
-def log_out_user(user_id)
-  if session[:user] == user_id
-    session[:user] = nil
-  end
-end
